% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{confounder.eg}
\alias{confounder.eg}
\title{confounder.eg}
\format{
An object of class \code{data.frame} with 50 rows and 5 columns.
}
\usage{
confounder.eg
}
\description{
Just test dataframe (A example table where the rows are samples and the columns are confounders.)
}
\examples{
  head(confounder.eg)
}
\keyword{datasets}
