% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groupInfo.eg}
\alias{groupInfo.eg}
\title{groupInfo.eg}
\format{
An object of class \code{data.frame} with 50 rows and 2 columns.
}
\usage{
groupInfo.eg
}
\description{
Just test dataframe (A example table where the rows are samples and the column is group information.)
}
\examples{
  head(groupInfo.eg)
}
\keyword{datasets}
