% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta2pathway.eg}
\alias{Meta2pathway.eg}
\title{Meta2pathway.eg}
\format{
An object of class \code{data.frame} with 30 rows and 45 columns.
}
\usage{
Meta2pathway.eg
}
\description{
Just test dataframe (A example table for Meta2pathway function.)
}
\examples{
  head(Meta2pathway.eg)
}
\keyword{datasets}
