% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabo.eg}
\alias{metabo.eg}
\title{metabo.eg}
\format{
An object of class \code{data.frame} with 50 rows and 10 columns.
}
\usage{
metabo.eg
}
\description{
Just test dataframe (A example table including metabolites with compound ID and predicted KEGG pathways from Meta2pathway function.)
}
\examples{
  head(metabo.eg)
}
\keyword{datasets}
