% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MicPathmatch.eg}
\alias{MicPathmatch.eg}
\title{MicPathmatch.eg}
\format{
An object of class \code{character} of length 30.
}
\usage{
MicPathmatch.eg
}
\description{
Just test dataframe (A example table for MicPathmatch function.)
}
\examples{
  head(MicPathmatch.eg)
}
\keyword{datasets}
