% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{micro.eg}
\alias{micro.eg}
\title{micro.eg}
\format{
An object of class \code{data.frame} with 50 rows and 10 columns.
}
\usage{
micro.eg
}
\description{
Just test dataframe (A example table including microbes and predicted KEGG pathways.)
}
\examples{
  head(micro.eg)
}
\keyword{datasets}
