% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMfunc.R
\name{MMfunc}
\alias{MMfunc}
\title{MMfunc}
\usage{
MMfunc(IIS, path_IISthre = 2.5, mm_IISthre = 2)
}
\arguments{
\item{IIS}{The result returned by IIS() function.}

\item{path_IISthre}{A IIS threshold value to identify the shared pathways.}

\item{mm_IISthre}{A IIS threshold value to identify the microbes/metabolites affiliation with the shared pathways.}
}
\value{
MMfunc_res A list including shared functional pathways.
}
\description{
A function is used to group microbes and metabolites in shared functional pathways based on different threshold setting according to functional pathway IIS,
microbes IIS and metabolites IIS.
}
\examples{
IIScore <- IIS(microApath = micro.eg, metaApath = metabo.eg,
               conf = confounder.eg, groupInfo = groupInfo.eg)
MMfunc_res <- MMfunc(IIS = IIScore)
}
