% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS.R
\name{IPS}
\alias{IPS}
\title{IPS}
\usage{
IPS(TarnetPairs, IFS)
}
\arguments{
\item{TarnetPairs}{The result returned by TarNet() function.}

\item{IFS}{The result returned by IFS() function.}
}
\value{
IPS_res A data frame including the importance pair score (IPS) of the target network.
}
\description{
A function is used to compute the importance pair score (IPS) of the network returned by the TarNet() function.
}
\examples{
IFScore <- IFS(microApath = micro.eg, metaApath = metabo.eg,
               conf = NULL, groupInfo = groupInfo.eg)
MMfunc_res <- MMfunc(IFS = IFScore)
TarNet_res <- TarNet(IFS = IFScore, mm2path = MMfunc_res)
IPScore <- IPS(TarnetPairs = TarNet_res,IFS = IFScore)
}
