% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMfunc.R
\name{MMfunc}
\alias{MMfunc}
\title{MMfunc}
\usage{
MMfunc(IFS, path_IFSthre = 2.5, mm_IFSthre = 2)
}
\arguments{
\item{IFS}{The result returned by IFS() function.}

\item{path_IFSthre}{A IFS threshold value to identify the shared pathways.}

\item{mm_IFSthre}{A IFS threshold value to identify the microbes/metabolites affiliation with the shared pathways.}
}
\value{
MMfunc_res A list including shared functional pathways.
}
\description{
A function is used to group microbes and metabolites in shared functional pathways based on different threshold setting according to functional pathway IFS,
microbes IFS and metabolites IFS.
}
\examples{
IFScore <- IFS(microApath = micro.eg, metaApath = metabo.eg,
               conf = NULL, groupInfo = groupInfo.eg)
MMfunc_res <- MMfunc(IFS = IFScore)
}
